Splits tests for functions from utils.py into test_utils.py
