+++ tests/scrape/test_scrape.py
@@ -13,23 +13,6 @@ from common.scrape_utils import (
 )
 
 
-alphabet = "abcdefghijklmnopqrstuvwxyz"
-months = [
-    "january",
-    "february",
-    "march",
-    "april",
-    "may",
-    "june",
-    "july",
-    "august",
-    "september",
-    "october",
-    "november",
-    "december",
-]
-
-
 class TestCheckAndParsePage:
     """Tests check_and_parse_page() which confirms that the current layout of
     the BOE page still matches the expected layout and returns a BeautifulSoup
@@ -140,29 +123,13 @@ class TestGetParseMeetingDate:
     """
 
     @pytest.mark.parametrize(
-        "word_a,word_b,expected_distance",
-        [
-            ("possible", "impossible", 2),
-            ("possible", "sorry", 7),
-            ("January", "Jaunary", 2),
-            ("June", "une", 1),
-        ],
-    )
-    def test_levenshtein_distance(self, word_a, word_b, expected_distance):
-        """Tests the Levenshtein distance algorithm created as a helper
-        function for parse_long_date"""
-
-        distance = levenshtein(word_a, word_b)
-        assert distance == expected_distance
-
-    @pytest.mark.parametrize(
         "input_date,output_date",
         [
             ("November 10, 2020", "2020_11_10"),  # checks standard date
             ("April 6, 2020", "2020_04_06"),  # checks for zero padding
             ("une 17, 2019", "2019_06_17"),  # checks for single letter deletion
-            ("Mrach 2, 2018", "2018_03_02"),
-        ],  # checks for swapped letters
+            ("Mrach 2, 2018", "2018_03_02"),  # checks for swapped letters
+        ],
     )
     def test_parse_long_dates(self, input_date, output_date):
         """Tests parse_long_dates() against the standard date format
@@ -290,58 +257,3 @@ class TestDownloadPDF:
           download the file specified by the link
         """
         assert 1
-
-
-class TestMonthSpellCheck:
-    """Tests the month misspelling detection.
-    This test also specifically *excludes* the misspellings of:
-      'juny'
-      'jule'
-    because both of these misspellings have Levenshtein distances
-    of 1 to both "june" and "july" and it's not possible to determine
-    the correct month without a lot of extra work. If we run into these
-    misspellings it will probably be easier to catch that specific error
-    than rework the function(s) to make the right call."""
-
-    def test_single_deletions(self):
-        """Test for correct detection of months with single-letter deletions."""
-        deletions = dict()
-        for month in months:
-            deletions[month] = list()
-            for j in range(len(month)):
-                deletions[month].append(month[:j] + month[j + 1 :])
-
-        for month, month_dels in deletions.items():
-            for deletion in month_dels:
-                match, score = levenshtein_match(deletion, MONTHS)
-                assert (
-                    match == month
-                ), f"month={month}, match={match}, score={score}, del={deletion}"
-
-    def test_single_misspell(self):
-        """Test for correct detection of months with single-letter changes.
-        Specifically exempts "jule" and "juny" as they are special cases that
-        hopefully never come up. (And if they do, it'll probably be easier
-        to specifically handle those errors than rework the spellchecking
-        to accomodate them)"""
-        misspellings = dict()
-        for month in months:
-            misspellings[month] = list()
-            for j in range(len(month)):
-                for char in alphabet:  # all possible single-letter changes
-                    misspell = month[:j] + char + month[j + 1 :]
-                    misspellings[month].append(misspell)
-
-        for month, month_misspells in misspellings.items():
-            for misspelling in month_misspells:
-                if misspelling in [
-                    "juny",
-                    "jule",
-                ]:  # specific exception for special case we hope to never see
-                    assert 1
-                else:
-                    match, score = levenshtein_match(misspelling, MONTHS)
-                    assert match == month, (
-                        f"month={month}, match={match}, score={score},"
-                        f"misspell={misspelling}"
-                    )
